Extract command.
